[TEST] Moved BulkProcessor tests to newly added BulkProcessorTests class
